+++ service/conversations/v1/service/configuration/api_op_client.go
@@ -1,7 +1,10 @@
 // Package configuration contains auto-generated files. DO NOT MODIFY
 package configuration
 
-import "github.com/RJPearson94/twilio-sdk-go/client"
+import (
+	"github.com/RJPearson94/twilio-sdk-go/client"
+	"github.com/RJPearson94/twilio-sdk-go/service/conversations/v1/service/configuration/notification"
+)
 
 // Client for managing service configuration
 // See https://www.twilio.com/docs/conversations/api/service-configuration-resource for more details
@@ -9,6 +12,8 @@ type Client struct {
 	client *client.Client
 
 	serviceSid string
+
+	Notification func() *notification.Client
 }
 
 // ClientProperties are the properties required to manage the configuration resources
@@ -22,5 +27,11 @@ func New(client *client.Client, properties ClientProperties) *Client {
 		client: client,
 
 		serviceSid: properties.ServiceSid,
+
+		Notification: func() *notification.Client {
+			return notification.New(client, notification.ClientProperties{
+				ServiceSid: properties.ServiceSid,
+			})
+		},
 	}
 }